Fix a test error for windows.

+++ k2/python/tests/fsa_test.py
@@ -682,8 +682,10 @@ class TestFsa(unittest.TestCase):
         fsa.symbols = symbols
         fsa.aux_symbols = aux_symbols
 
-        fsa.draw(filename='foo.png')
-        os.remove('foo.png')
+        import shutil
+        if shutil.which('dot') is not None:
+          fsa.draw(filename='foo.png')
+          os.remove('foo.png')
 
     def test_to(self):
         s = '''
